requirement.txt
transformers
transformers[torch]
Pillow
grpcio
grpcio-tools
